+++ R/animateSpectra.R
@@ -21,25 +21,26 @@
 #' to logarithmic size bins.
 #' @param total A boolean value that determines whether the total over all
 #'   species in the system is plotted as well. Default is FALSE.
+#' @param resource Deprecated compatibility argument from mizer's generic.
+#' @param background Deprecated compatibility argument from mizer's generic.
+#' @param ... Other arguments (currently unused).
 #'
 #' @return A plotly object
-#' @export
 #' @family plotting functions
-animateSpectra <- function(sim,
-                           species = NULL,
-                           resources = NULL,
-                           time_range,
-                           wlim = c(NA, NA),
-                           ylim = c(NA, NA),
-                           power = 1,
-                           total = FALSE) {
-    if (is.null(getComponent(getParams(sim), "MR"))) {
-        return(mizer::animateSpectra(sim = sim, species = species,
-                                     time_range = time_range,
-                                     wlim = wlim, ylim = ylim,
-                                     power = power, total = total,
-                                     resource = TRUE))
-    }
+#' @export
+#' @name animateSpectra
+animateSpectra.mizerMRSim <- function(sim,
+                                      species = NULL,
+                                      time_range,
+                                      wlim = c(NA, NA),
+                                      ylim = c(NA, NA),
+                                      power = 1,
+                                      total = FALSE,
+                                      resource = TRUE,
+                                      background = TRUE,
+                                      ...,
+                                      resources = NULL) {
+    NextMethod(resource = FALSE, background = background, ...)
     species <- valid_species_arg(sim, species)
     resources <- valid_resources_arg(sim, resources)
     if (missing(time_range)) {